Create DBAccess.cs
DBAccess class of the Data Access Layer

By: Philip McCarter

What does this class do?
This class creates the access to the database, as well as provide the direct interaction with the database. Currently, only one method is coded, which is a method for reading data from the database. It works as follows:
•	The method takes a table name and a stored procedure name as parameters
•	The database connection is opened
•	An SqlDataReader initiated, using the stored procedure name and connection name
•	The SqlDataReader then begins to read the data, and stores what is read into a DataTable
•	Once the reader is finished, the connection is closed.
•	There is an exception in the event that the database cannot be accessed
•	The database is returned at the end of the method.
This class may not work currently, as it requires minor tweaks in its functionality to match the stored procedure from the database.